+++ CHANGELOG.md
@@ -1,12 +1,31 @@
 Changelog
 =========
 
+v0.2
+-----------
+
+* Fixed Group Discount doesn't work correctly
+* Added Permission `RepairRecipe.repair.overRepair`
+* Fixed Enchantment Multiplier could reduce repair costs
+* Added Enchantment Multiplier per Group Configuration
+
+
+v0.2-alpha1
+-----------
+
+* Added Permissions `RepairRecipe.repair` and `RepairRecipe.repair.enchant`
+* Added OverRepair Configuration
+* Added Group Discount Configuration
+* Added Keep Enchantments Configuration
+* Added Enchantment Multiplier Configuration
+
+
 v0.1.1
 ------
 
-Added Shears
+* Added Shears
 
 v0.1
 ----
 
-Initial Commit
+* Initial Commit